requirements.txt
annotated-types
anyio
asgiref
# backports.zoneinfo
# bcml
boto3
botocore
# botw-utils
cefpython3
certifi
cffi
charset-normalizer
clr_loader
cryptography
deprecation
Django
django-storages
exceptiongroup
gotrue
h11
h2
hpack
httpcore
httpx
hyperframe
idna
jmespath
# oead
packaging
pillow
postgrest
proxy_tools
pycparser
pydantic
pydantic_core
pyparsing
python-dateutil
python-dotenv
# pythonnet
pywebview
realtime
requests
# rstb
s3transfer
six
sniffio
sqlparse
storage3
StrEnum
supabase
supafunc
typing_extensions
tzdata
urllib3
websockets
xxhash
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
gunicorn
alembic
posthog


